Introduce wrapper for serve and run
We are coming to the limits of what we can do in a "podman run"
line. Create wrapper functions so we can do things like forking
processes and other similar things that you need to do inside a
container in python3. There are some features coming up where
rather than upstreaming separate solutions to all our engines
like vLLM and llama.cpp we want to solve the problem in the
python3 layer.

The "if True:"'s will remain for a while, we may need to wait for
containers to be distributed around the place before we turn things
on.

Signed-off-by: Eric Curtin <ecurtin@redhat.com>